# docs/swagger.yaml
basePath: /bookcabin
definitions:
  github_com_elkoshar_bookcabin_pkg_response.Error:
    properties:
      code:
        example: 0
        type: integer
      msg:
        example: ' '
        type: string
      status:
        example: false
        type: boolean
    type: object
  github_com_elkoshar_bookcabin_pkg_response.Response:
    properties:
      code: {}
      data: {}
      error:
        $ref: '#/definitions/github_com_elkoshar_bookcabin_pkg_response.Error'
      message:
        type: string
      serverTime:
        type: integer
    type: object
  service.AirlineInfo:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  service.DurationInfo:
    properties:
      formatted:
        type: string
      total_minutes:
        type: integer
    type: object
  service.LocationInfo:
    properties:
      airport:
        type: string
      city:
        type: string
      datetime:
        type: string
      timestamp:
        type: integer
    type: object
  service.Metadata:
    properties:
      providers_failed:
        type: integer
      providers_queried:
        type: integer
      providers_succeeded:
        type: integer
      search_time_ms:
        type: integer
      total_results:
        type: integer
    type: object
  service.PriceInfo:
    properties:
      amount:
        type: number
      currency:
        type: string
      formatted:
        type: string
    type: object
  service.SearchCriteria:
    properties:
      cabinClass:
        type: string
      departureDate:
        type: string
      destination:
        type: string
      origin:
        type: string
      passengers:
        type: integer
      returnDate:
        type: string
    type: object
  service.SearchResponse:
    properties:
      flights:
        items:
          $ref: '#/definitions/service.UnifiedFlight'
        type: array
      metadata:
        $ref: '#/definitions/service.Metadata'
      search_criteria:
        $ref: '#/definitions/service.SearchCriteria'
    type: object
  service.UnifiedFlight:
    properties:
      airline:
        $ref: '#/definitions/service.AirlineInfo'
      amenities:
        items:
          type: string
        type: array
      arrival:
        $ref: '#/definitions/service.LocationInfo'
      available_seats:
        type: integer
      cabin_class:
        type: string
      departure:
        $ref: '#/definitions/service.LocationInfo'
      duration:
        $ref: '#/definitions/service.DurationInfo'
      flight_number:
        type: string
      id:
        type: string
      price:
        $ref: '#/definitions/service.PriceInfo'
      provider:
        type: string
      stops:
        type: integer
    type: object
info:
  contact:
    email: elko.s.eppasa@mail.com
    name: Elko Sharhadi Eppasa
    url: https://github.com/elkoshar
  description: This service is to handle Flight Search and Aggregation API. For more
    detail, please visit https://github.com/elkoshar/bookcabin
  title: Flight Search and Aggregation API
  version: "0.1"
paths:
  /flight/search:
    post:
      consumes:
      - application/json
      description: SearchFlight handles request for searching flights
      parameters:
      - default: id
        description: accept language
        in: header
        name: Accept-Language
        required: true
        type: string
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.SearchCriteria'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_elkoshar_bookcabin_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SearchResponse'
              type: object
      summary: Search Flight
      tags:
      - Flight
swagger: "2.0"